data-pipeline/requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
kafka-python==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dateutil==2.8.2
